fix: Bulk select doesn't properly add data to rows (#1867)
When a user selects a CVE and then selects systems one-by-one, the `selectedRows` array in `entities` contains all of the data needed for the remediation wizard.

When a user bulk selects any number of systems, the `selectedRows` array in `entities` does not include all of the necessary data. This leaves out critical `rules` data needed to provide the proper call to the resolution api.

Steps to Reproduce:
Go to vulnerabilities -> CVEs
Click the link for a CVE (make sure it's a CVE with multiple resolutions)
Select systems one-by-one and check the entities.selectedRows array, it will have the full data for each row
Now, bulk select any number of systems and check the entities.selectedRows array again. it will have much less data

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>