fix(RHINENG-7963): Bug with select all CVEs option on CVEs page
If a user has Edge systems in their account, we need to convert the affecting param to be in affectingHostType
So, we use hybrid system filter to determine this.

- If the user have both Conventional & Edge systems, we convert the affecting params to be in affectingHostType params
because affecting is no longer being used in the backend server in this case.
- If the user has only Conventional systems, the backend server still uses affecting param, so there is no need for the conversion.

Because affecting param is no longer being used in the backend server in the first case, sending it in the request would make the request fail, or in case the request does not fail, it would return bad results

Link to issue: https://issues.redhat.com/browse/RHINENG-7963